move web.js to src
